API - fix too delete pics only after email sent is success

+++ src/main/java/com/squirrel/persons/job/VisitorsJob.java
@@ -32,10 +32,11 @@ public class VisitorsJob {
     @Scheduled(fixedDelay = 60000)
     public void triggerJob() throws MessagingException, DocumentException, IOException {
         LOGGER.info("Triggering unknown visitors job");
-        emailService.attachImagesAndSendEmail(toEmailAddress, VISITOR_PATH,"Unknown visitors",
-                "People who were near your property today");
-        FileUtils.copyAllFiles(VISITOR_PATH, VISITOR_ARCHIVE_PATH);
-        FileUtils.deleteImages(VISITOR_ARCHIVE_PATH);
+        if(emailService.attachImagesAndSendEmail(toEmailAddress, VISITOR_PATH,"Unknown visitors",
+                "People who were near your property today")) {
+            FileUtils.copyAllFiles(VISITOR_PATH, VISITOR_ARCHIVE_PATH);
+            FileUtils.deleteImages(VISITOR_ARCHIVE_PATH);
+        }
     }
 
 }
